# academy_clean/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for أكاديمية البرمجة المتكاملة
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users with ranking
      parameters:
        - name: role
          in: query
          schema:
            type: string
            enum: [user, creator, employer, admin]
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserListResponse"
    post:
      operationId: createUser
      tags: [users]
      summary: Create or upsert a user profile (called after Clerk auth)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserBody"
      responses:
        "201":
          description: Created user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{userId}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get a user profile by ID
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user profile
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/leaderboard:
    get:
      operationId: getLeaderboard
      tags: [users]
      summary: Get top-ranked users leaderboard
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Leaderboard
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LeaderboardEntry"

  # ─── Courses ─────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
        clerkId:
          type: string
        name:
          type: string
        username:
          type: string
        email:
          type: string
        avatarUrl:
          type: string
        bio:
          type: string
        role:
          type: string
          enum: [user, creator, employer, admin]
        points:
          type: integer
        globalRank:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, clerkId, name, username, email, role, points, createdAt]
    UserProfile:
      allOf:
        - $ref: "#/components/schemas/User"
        - type: object
          properties:
            solvedChallenges:
              type: integer
            totalCourses:
              type: integer
            totalRepositories:
              type: integer
            challengeCategories:
              type: array
              items:
                type: object
                properties:
                  category:
                    type: string
                  count:
                    type: integer
                required: [category, count]
          required: [solvedChallenges, totalCourses, totalRepositories, challengeCategories]
    UserListResponse:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total:
          type: integer
      required: [users, total]
    LeaderboardEntry:
      type: object
      properties:
        rank:
          type: integer
        user:
          $ref: "#/components/schemas/User"
        points:
          type: integer
        solvedChallenges:
          type: integer
      required: [rank, user, points, solvedChallenges]
    CreateUserBody:
      type: object
      properties:
        clerkId:
          type: string
        name:
          type: string
        username:
          type: string
        email:
          type: string
        avatarUrl:
          type: string
        role:
          type: string
          enum: [user, creator, employer, admin]
          default: user
      required: [clerkId, name, username, email]
    UpdateUserBody:
      type: object
      properties:
        name:
          type: string
        bio:
          type: string
        avatarUrl:
          type: string

    # ─── Course schemas ───────────────────────────────────────────────────────
...